+++ sc/social/like/config.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
-# File: config.py
-
-__author__ = """Simples Consultoria <products@simplesconsultoria.com.br>"""
-__docformat__ = 'plaintext'
-
 PROJECTNAME = 'sc.social.like'
 
 DEFAULT_ENABLED_CONTENT_TYPES = ('Document', 'Event', 'News Item')
 DEFAULT_PLUGINS_ENABLED = ('Facebook', 'Twitter')
+
+TILES = [
+    u'sc.social.like.facebook',
+    u'sc.social.like.twitter',
+]