+++ sc/social/like/testing.py
@@ -8,6 +8,14 @@ from plone.app.testing import PloneSandboxLayer
 from plone.testing import z2
 
 import os.path
+import pkg_resources
+
+try:
+    pkg_resources.get_distribution('collective.cover')
+except pkg_resources.DistributionNotFound:
+    HAS_COVER = False
+else:
+    HAS_COVER = True
 
 IS_PLONE_5 = api.env.plone_version().startswith('5')
 
@@ -25,15 +33,21 @@ class Fixture(PloneSandboxLayer):
     defaultBases = (PLONE_FIXTURE,)
 
     def setUpZope(self, app, configurationContext):
-        # Load ZCML
+        if HAS_COVER:
+            import collective.cover
+            self.loadZCML(package=collective.cover)
+
         import sc.social.like
         self.loadZCML(package=sc.social.like)
 
     def setUpPloneSite(self, portal):
-        # Install into Plone site using portal_setup
+        if HAS_COVER:
+            self.applyProfile(portal, 'collective.cover:default')
+
         self.applyProfile(portal, 'sc.social.like:default')
 
 FIXTURE = Fixture()
+
 INTEGRATION_TESTING = IntegrationTesting(
     bases=(FIXTURE,),
     name='sc.social.like:Integration',